make preview table more reuseable
